v0.3.2 piggo strike
