+++ core/main.ts
@@ -18,7 +18,7 @@ export * from "./src/contrib/systems/net/WsClientSystem";
 export * from "./src/contrib/systems/net/WsServerSystem";
 export * from "./src/contrib/systems/ui/ClickableSystem";
 export * from "./src/contrib/systems/ui/DebugSystem";
-export * from "./src/contrib/systems/ui/GuiSystem";
+export * from "./src/contrib/systems/ui/HealthBarSystem";
 export * from "./src/contrib/systems/ui/InputSystem";
 export * from "./src/contrib/systems/ui/RenderSystem";
 export * from "./src/contrib/entities/buttons/ConnectButton";
@@ -31,7 +31,8 @@ export * from "./src/contrib/entities/core/Noob";
 export * from "./src/contrib/entities/core/SystemEntity";
 export * from "./src/contrib/entities/objects/Ball";
 export * from "./src/contrib/entities/objects/Goal";
-export * from "./src/contrib/entities/terrain/SpaceBackground";
+export * from "./src/contrib/entities/objects/Projectile";
+export * from "./src/contrib/entities/terrain/Background";
 export * from "./src/contrib/entities/terrain/TileFloor";
 export * from "./src/contrib/entities/terrain/TilingSpriteFloor";
 export * from "./src/contrib/entities/terrain/Wall";
