+++ core/src/contrib/components/renderables/HealthBar.ts
@@ -11,12 +11,12 @@ export const HealthBar = (health: Health): (r: Renderable) => Promise<void> => {
   const draw = (g: Graphics) => {
     // gold outline
     g.setStrokeStyle({ width: 1, color: 0xffd700, alpha: 1 });
-    g.rect(-15, -35, 30, 5);
+    g.rect(-15, -30, 30, 5);
     g.stroke();
 
     // red length proportional to percent health
     const length = 30 * (health.data.health / health.data.maxHealth);
-    g.rect(-15, -35, length, 5);
+    g.rect(-15, -30, length, 5);
     g.fill({ color: 0xff0000, alpha: 0.9 });
   }
