+++ core/src/contrib/entities/ui/Chat.ts
@@ -1,5 +1,4 @@
-import { Entity, World } from "@piggo-gg/core";
-import { Position, Renderable, TextBox, chatBuffer, chatIsOpen } from "@piggo-gg/core";
+import { Entity, Position, Renderable, TextBox, World, chatBuffer, chatIsOpen } from "@piggo-gg/core";
 import { Text } from "pixi.js";
 
 export const Chat = (): Entity => {
