+++ core/src/contrib/entities/ui/FpsText.ts
@@ -1,5 +1,4 @@
-import { Entity, World } from "@piggo-gg/core";
-import { Position, Renderable } from "@piggo-gg/core";
+import { Entity, Position, Renderable, World } from "@piggo-gg/core";
 import { Text } from "pixi.js";
 
 export type FpsTextProps = {
@@ -15,7 +14,7 @@ export const FpsText = ({ x, y }: FpsTextProps = {}) => Entity<Position | Render
     }),
     renderable: new Renderable({
       zIndex: 3,
-      container: async () => new Text("", { fontSize: 14, fill: "#FFFFFF" }),
+      container: async () => new Text({ text: "", style: { fontSize: 14, fill: "#FFFFFF" } }),
       dynamic: (t: Text, _, __, w: World) => {
         if (w.tick % 5 !== 0) return;
         if (t) {
@@ -36,7 +35,7 @@ export const LagText = ({ x, y }: FpsTextProps = {}) => Entity<Position | Render
     }),
     renderable: new Renderable({
       zIndex: 3,
-      container: async () => new Text("", { fontSize: 14, fill: "#FFFFFF" }),
+      container: async () => new Text({ text: "", style: { fontSize: 14, fill: "#FFFFFF" } }),
       dynamic: (t: Text, _, __, w: World) => {
         const lag = Math.round(w.ms);
         if (w.tick % 5 !== 0) return;
