+++ core/src/ecs/System.ts
@@ -8,7 +8,7 @@ export interface System<T extends string = string> {
   query?: string[];
   skipOnRollback?: boolean
   onRender?: (entities: Entity[]) => void;
-  onRemove?: () => void;
+  onCleanup?: () => void;
   onRollback?: () => void
   onTick: (entities: Entity[], isRollback: boolean) => void;
 }
