+++ core/src/runtime/Game.ts
@@ -1,4 +1,4 @@
-import { Entity, System, SystemBuilder, World } from "@piggo-gg/core";
+import { Entity, SystemBuilder, World } from "@piggo-gg/core";
 
 // a game is a collection of entities and systems
 export type Game<T extends string = string> = {
