+++ games/main.ts
@@ -1,4 +1,5 @@
+export * from "./legends/Legends";
+export * from "./legends/entities";
 export * from "./soccer/Soccer";
 export * from "./soccer/entities";
-export * from "./piggo-legends/PiggoLegends";
-export * from "./piggo-legends/entities";
+export * from "./strike/Strike";
